requirements.txt
matplotlib
opencv-python
numpy
Pillow